% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assoc.R
\name{plot_assoc}
\alias{plot_assoc}
\title{Makes a table with associations}
\usage{
plot_assoc(model_output = NA, model = NA, indep.var)
}
\arguments{
\item{model_output}{The output from mplus model}

\item{model}{The model extracted using `MplusAutomation::readModel`}
}
\value{
A table that contains the parameter estimate, standard error and p-value
for each `ON` relationship.
}
\description{
`plot_assoc` takes the parameter estimate from the model output
that are associated with 'ON' operator.
}
