% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot correlation matrix}
\usage{
plot_cor(model_output)
}
\arguments{
\item{model_output}{The output from mplus model}
}
\description{
Plot correlation matrix
}
\author{
Subrata Paul
}
