% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esem.R
\name{plot_esem}
\alias{plot_esem}
\title{Plots parameters in an ESEM model}
\usage{
plot_esem(
  model_output,
  covariates = NA,
  indep.var = NA,
  load.thres = 0.2,
  sort = F
)
}
\arguments{
\item{model_output}{The output from MPLUS}

\item{covariates}{A character vector of covariates. If the dependent variables were
regressed on some variables they are the covariates.}

\item{indep.var}{The variables that were regressed on the factors}

\item{load.thres}{Threshold of loading to include in the plot. Default is 0.2}

\item{sort}{Logical. If true the loadings will be sorted using `fa.sort` function}
}
\description{
Plots parameters in an ESEM model
}
\examples{
{
mplus_out = system.file("extdata", "sample_esem.out", package = "PlotMplus")
plot_esem(model_output = mplus_out, covariates = c('AGE','SEX','BMI'),indep.var = 'CREACT',sort = T)
}
}
\author{
Subrata Paul
}
