% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esem_marginal.R
\name{plot_esem_marginal}
\alias{plot_esem_marginal}
\title{Plots parameters in an ESEM model}
\usage{
plot_esem_marginal(
  model_output,
  marginal,
  marginal_show,
  covariates = NA,
  indep.var = NA,
  load.thres = 0.2,
  sort = F,
  load.text.size = 5
)
}
\arguments{
\item{model_output}{The output from MPLUS}

\item{marginal}{Dataframe. Contains marginal association of the independent variable with the dependents.
The dataframe contains columns with names "Condition", "Estimate", "SE", "Tstat", "Pvalue", "R2" , "Tstat_times_R2"}

\item{marginal_show}{Either `Estimate` or `Tstat_times_R2`. The selected statistic will be showed.}

\item{covariates}{A character vector of covariates. If the dependent variables were
regressed on some variables they are the covariates.}

\item{indep.var}{The variables that were regressed on the factors}

\item{load.thres}{Threshold of loading to include in the plot. Default is 0.2}

\item{sort}{Logical. If true the loadings will be sorted using `fa.sort` function}
}
\description{
Plots parameters in an ESEM model
}
\examples{
{
mplus_out = system.file("extdata", "sample_esem.out", package = "PlotMplus")
plot_esem(model_output = mplus_out, covariates = c('AGE','SEX','BMI'),indep.var = 'CREACT',sort = T)
}
}
\author{
Subrata Paul
}
