% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_cor.R
\name{plot_factor_cor}
\alias{plot_factor_cor}
\title{Plots correlation matrix of the factors}
\usage{
plot_factor_cor(model_output)
}
\arguments{
\item{model_output}{Output from Mplus model}
}
\value{
Correlation plot of factors
}
\description{
Plots correlation matrix of the factors
}
