% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plots the loading matrix}
\usage{
plot_loadings(model_output, load.thres = NA, sort = T)
}
\arguments{
\item{model_output}{Output from mplus run}

\item{load.thres}{Threshold of loading to avoid plotting. Default is to plot all laoding.}

\item{sort}{Logical. If true the variables will be sorted using `fa.sort` function.}
}
\description{
Given the mplus output the function extract the loading matrix
and make a box plot.
}
\author{
Subrata Paul
}
